+++ src/survey/views/survey.py
@@ -40,9 +40,9 @@ class SurveyView(APIView):
                     "error_title": "Session expired",
                     "error_message": f"Your session is expired, log ",
                     "add_link": True,
-                    "link_url": f"{os.getenv('INSTANCE')}/hub/frontend/oauth/login?client_id={os.getenv('APP_ID')}",
+                    "link_url": f"{os.getenv('WENET_INSTANCE')}/hub/frontend/oauth/login?client_id={os.getenv('APP_ID')}",
                     "link_text": "here"
                 }
                 return render(request, 'authentication/error.html', context=context)
         else:
-            return redirect(os.getenv("BASE_PATH"))
+            return redirect(os.getenv("BASE_URL"))