+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
-## 0.*
+## NEXT
+
+* Set up the project.
+* Added oauth flow with wenet.
+* Created basic HTML pages.